// src/generated/client/worldmonitor/economic/v1/service_client.ts
export interface BisExchangeRate {
  countryCode: string;
  countryName: string;
  realEer: number;
  nominalEer: number;
  realChange: number;
  date: string;
}